ruta delete creada
