+++ api/routes/index.js
@@ -1,9 +1,8 @@
 const express = require('express')
-const { get } = require('../controllers/index')
-
+const productRoutes = require("./productRoutes.js")
 const router = express.Router()
 
-// example of a route with index controller get function
-router.get('/', get)
+
+router.use("/products", productRoutes)
 
 module.exports = router
