+++ api/routes/productRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const { deleteProduct } = require('../controllers/Product/deleteProduct.js')
+
+const router = express.Router()
+
+router.delete('/:productId', deleteProduct)
+
+module.exports = router
